+++ eodag_cube/api/product/__init__.py
@@ -16,4 +16,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """EODAG product package"""
+from ._assets import Asset, AssetsDict  # noqa
 from ._product import EOProduct  # noqa